Split execution acknowledgement source mapping
